requirements.txt
pandas>=1.3.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0 